requirements.txt
fastapi
uvicorn
pydantic
aiohttp
supabase
typing-extensions
azure_cognitiveservices-speech
ffmpeg
google-cloud-pubsub
pytest
httpx
pytest-asyncio
pytest-mock
pytest-cov
numpy
python-dotenv
wave